+++ src/engine/GroupingAlgorithms.java
@@ -5,19 +5,26 @@ import java.util.LinkedList;
 
 import matchmaker.PersonalityRule;
 
-public class GroupingRunner {
-	public static void main(String[] args) {
-		PersonalityRule.ruleList = new LinkedList<>();
+
+
+public class GroupingAlgorithms {
+	static LinkedList<Student> list;
+	
+	public static void initialize() {
 		CSVLoader dataLoader;
 		RuleLoader ruleLoader;
-		LinkedList<Student> list;
 		
+		
+		PersonalityRule.ruleList = new LinkedList<>();
 		dataLoader = new CSVLoader(new File("C:\\Users\\Thai Nguyen\\Desktop\\Form1.csv"));
 		ruleLoader = new RuleLoader(new File("C:\\Users\\Thai Nguyen\\Desktop\\Rule1.csv"));
 		
 		list = dataLoader.readGoogleFile();
-		ruleLoader.generateRulesFromProgram(); //WTF are you doing here?
 		ruleLoader.readFile();
+	}
+	
+	public static void main(String[] args) {
+		initialize();
 		
 		for (Student s : list) {
 			System.out.println(s);
@@ -26,12 +33,10 @@ public class GroupingRunner {
 		
 		for (int i = 0; i < list.size(); i++) {
 			Student a = list.get(i);
-			Student b;
 			for (int j = i + 1; j < list.size(); j++) {
-				if (!a.equals((b = list.get(j)))) {
-					System.out.print(String.format("%+3d", a.matchWith(b)));
-					System.out.println("\t" + a + " / " + b);
-				}
+				Student b = list.get(j);
+				System.out.print(String.format("%+3d", a.matchWith(b)));
+				System.out.println("\t" + a + " / " + b);
 			}
 		}
 	}
